+++ src/blueprint/mod.rs
@@ -1,9 +1,11 @@
 mod lang;
+mod context;
 mod reader;
 mod renderer;
 mod store;
 mod syntax;
 
+pub(crate) use context::*;
 pub use lang::*;
 pub use reader::*;
 pub use renderer::*;
